send fish coords in percentages to view

+++ server/src/fishes.h
@@ -9,7 +9,7 @@ struct fish_t {
     float current_x, current_y;
     float width, height;
     float target_x, target_y;
-    float time_left;
+    int time_left;
     TAILQ_ENTRY(fish_t) _next;
 };
